+++ GPTPullRequestReview/src/getFileExtension.ts
@@ -1,0 +1,3 @@
+export function getFileExtension(fileName: string) {
+  return fileName.slice(((fileName.lastIndexOf(".") - 1) >>> 0) + 2);
+}
