+++ GPTPullRequestReview/src/review.ts
@@ -57,8 +57,4 @@ async function run() {
   }
 }
 
-export function getFileExtension(fileName: string) {
-  return fileName.slice(((fileName.lastIndexOf(".") - 1) >>> 0) + 2);
-}
-
 run();
