+++ GPTPullRequestReview/tsconfig.json
@@ -9,5 +9,8 @@
     "outDir": "./dist",                        /* Redirect output structure to the directory. */
     "rootDir": "./src",                        /* Specify the root directory of input files. Use to control the output directory structure with --outDir. */
     "moduleResolution": "node",
+    "paths": {
+      "@/*": ["./src/*"]
+    }
   }
 }
